[exp2/*] Include experiment 2 sources
This experiment consists in modularize the code written in experiment 1.
A abstraction layer was made for the RGB LED module, so that the
functions that include this header only have to know which color it
wants to control.
The delay function was also separated in a different source. Further
improvements will be made in order to relate absolute time and
instruction execution (loop delays)